Refactor document validation into cohesive services
